Feature Request
Hey, as Issues are not allowed (is that on purpose?), I wanted to propose the following feature, but maybe it is out of scope:

# Problem
I have two headphones:
One of them can only connect to one device at a time. That one I would like to disconnect on snooze.
The other one can connect to two devices at a time. That one I would like not to disconnect on snooze.

# Possible Solution
I am aware that blueznooze currently totally disables bluetooth. This feature would require to selectively disconnect devices

# Alternatives
It might be possible to achieve this with another tool, for example http://www.hammerspoon.org/